backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg[binary]>=3.2.0
stripe==8.0.0
python-decouple==3.8
Pillow>=10.4.0
django-extensions==3.2.3
qrcode[pil]==7.4.2
